requirements.txt
absl-py==2.1.0
anyio==3.7.1
astor==0.8.1
attrs==23.2.0
autolab-core==1.1.1
autolab-perception==1.0.0
blas==1.0
ca-certificates==2024.11.26
certifi==2022.12.7
chardet==5.2.0
colorama==0.4.6
colorlog==6.8.2
cudatoolkit==10.2.89
cycler==0.11.0
dill==0.3.7
embreex==2.17.7.post5
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
fonttools==4.38.0
freetype==2.4.10
freetype-py==2.4.0
future==1.0.0
gast==0.2.2
google-pasta==0.2.0
gputil==1.4.0
grpcio==1.62.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
icc_rt==2022.1.0
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
intel-openmp==2023.1.0
joblib==1.3.2
jpeg==8d
jsonschema==4.17.3
keras-applications==1.0.8
keras-preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.5
libpng==1.6.17
libtiff==4.0.2
libuv==1.48.0
libwebp==1.2.4
libwebp-base==1.2.4
lxml==5.2.2
mapbox-earcut==1.0.1
markdown==3.4.4
markupsafe==2.1.5
matplotlib==3.5.3
mkl==2020.2
mkl_fft==1.0.12
mkl_random==1.1.1
msys2-conda-epoch==20160418
multiprocess==0.70.15
networkx==2.6.3
ninja==1.7.2
numpy==1.20.0
opencv-python==4.10.0.84
openssl==1.1.1w
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
pillow==9.5.0
pip==22.3.1
pkgutil-resolve-name==1.3.10
protobuf==3.20.0
psutil==6.0.0
pycollada==0.8
pyglet==1.5.21
pyopengl==3.1.0
pyparsing==3.1.2
pyreadline==2.1
pyrealsense2==2.55.1.6486
pyrender==0.1.45
pyrsistent==0.19.3
pyserial==3.5
python==3.7.16
python-dateutil==2.9.0.post0
pytorch-mutex==1.0
pytz==2025.1
pywavelets==1.3.0
rtree==1.0.1
ruamel-yaml==0.17.32
ruamel-yaml-clib==0.2.7
scienceplots==2.1.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setproctitle==1.3.3
setuptools==65.6.3
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
sqlite==3.45.3
svg-path==6.2
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tk==8.6.12
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
trimesh==4.4.1
typing-extensions==4.7.1
typing_extensions==4.3.0
vc==14.2
vhacdx==0.0.8.post1
visualization==1.0.0
vs2015_runtime==14.29.30133
werkzeug==2.2.3
wheel==0.38.4
wincertstore==0.2
wrapt==1.16.0
xatlas==0.0.9
xxhash==3.4.1
zipp==3.15.0
zlib==1.2.8
